feat: implement full item/inventory/shop/crafting system (Phase 5)
Complete item economy for Chronicles of Aetheria:

Item Database (128 items across 5 categories):
- Equipment: 34 items (common → legendary, mirrors char_equip.h)
- Common consumables: potions, ethers, antidotes, buffs, grimoires
- Rare consumables: elixirs, soma, phoenix, megalixir, legendary items
- Crafting materials: 40 items (iron ore → dragon scale / astral dust)
- Key items: 10 story items (dungeon keys, codex, aetheria key)

Party Inventory:
- Shared 80-slot bag with stacking (up to 99 per stack)
- Gold management (max 999999) with spend/add guards
- Compact + sort utilities, serialize/deserialize for save

Item Use System:
- Field and battle context resolution
- 20 consumable effect types (CEFF_HEAL_HP through CEFF_XP_BOOST)
- equip_from_bag / unequip_to_bag with auto-recompute
- Party-wide vs single-target logic

Crafting System:
- 24 recipes (BASIC → FORGE → ALCHEMY → LEGENDARY)
- Discovery bitmask (known_default = 3 starter recipes)
- Ingredient check + consume + produce in one transaction
- Legendary recipes require rare boss-drop materials

Shop System:
- 8 shops (village, weapons, armor, magic, guild, black market,
  legendary forge, alchemist)
- Buy/sell with per-shop price multipliers (90%-150% buy, 40%-60% sell)
- Access requirements (guild pass, hero badge, aetheria key)
- Full gold integration

Infrastructure:
- EVT_LEVEL_UP, EVT_ITEM_USED, EVT_SHOP_BUY, EVT_SHOP_SELL added
- char_equip_attr_bonus type fix (s8[] not s16[])
- Makefile updated for game/item/ source tree

+++ chronicles-of-aetheria/src/game/character/character.c
@@ -75,10 +75,10 @@ void character_recompute(Character* ch)
 
     /* Add equipment attribute bonuses on top */
     {
-        s16 attr_eq[ATTR_COUNT];
+        s8 attr_eq[ATTR_COUNT];
         equip_compute_attr_bonus(ch, attr_eq);
         for (i = 0; i < ATTR_COUNT; i++)
-            ch->base.bonus[i] += attr_eq[i];
+            ch->base.bonus[i] += (s16)attr_eq[i];
     }
 
     stats_compute_derived(&ch->base, &ch->derived, ch->cls,
